Further improve modularization/factoring of type system stuff
See the grandparent commit's message for rationale.
