+++ src/language/semantics/type-system/type-formats/type.ts
@@ -1,13 +1,11 @@
-import type {
-  ApplicationType,
-  FunctionType,
-  IndexedAccessType,
-  IntrinsicApplicationType,
-  ObjectType,
-  TypeParameter,
-  UnionType,
-} from '../type-formats.js'
+import type { ApplicationType } from './application-type.js'
+import type { FunctionType } from './function-type.js'
+import type { IndexedAccessType } from './indexed-access-type.js'
+import type { IntrinsicApplicationType } from './intrinsic-application-type.js'
+import type { ObjectType } from './object-type.js'
 import type { OpaqueType } from './opaque-type.js'
+import type { TypeParameter } from './type-parameter-type.js'
+import type { UnionType } from './union-type.js'
 
 export type Type =
   | ApplicationType
