+++ src/language/semantics/type-system/type-parameter-analysis.ts
@@ -1,12 +1,9 @@
 import { something } from './prelude-types.js'
 import { simplifyUnionType } from './subtyping.js'
-import {
-  makeUnionType,
-  matchTypeFormat,
-  type Type,
-  type TypeParameter,
-  type UnionType,
-} from './type-formats.js'
+import { matchTypeFormat } from './type-formats/match-type-format.js'
+import { type TypeParameter } from './type-formats/type-parameter-type.js'
+import type { Type } from './type-formats/type.js'
+import { makeUnionType, type UnionType } from './type-formats/union-type.js'
 import {
   functionParameterKey,
   functionReturnKey,
